requirements.txt
BeautifulSoup==3.2.1
django-extensions==0.8